+++ src/Integration/FatalErrorListenerIntegration.php
@@ -17,17 +17,21 @@ use Sentry\State\Hub;
 final class FatalErrorListenerIntegration implements IntegrationInterface
 {
     /**
-     * @var Options The options, to know which error level to use
+     * @var Options|null The options, to know which error level to use
      */
     private $options;
 
     /**
      * Constructor.
      *
-     * @param Options $options The options to be used with this integration
+     * @param Options|null $options The options to be used with this integration
      */
-    public function __construct(Options $options)
+    public function __construct(?Options $options = null)
     {
+        if (null !== $options) {
+            @trigger_error(sprintf('Passing the options as argument of the constructor of the "%s" class is deprecated since version 2.1 and will not work in 3.0.', self::class), E_USER_DEPRECATED);
+        }
+
         $this->options = $options;
     }
 
@@ -38,11 +42,30 @@ final class FatalErrorListenerIntegration implements IntegrationInterface
     {
         $errorHandler = ErrorHandler::registerOnceFatalErrorHandler();
         $errorHandler->addFatalErrorHandlerListener(function (FatalErrorException $exception): void {
-            if (!($this->options->getErrorTypes() & $exception->getSeverity())) {
+            $currentHub = Hub::getCurrent();
+            $client = $currentHub->getClient();
+
+            // The client could have been detached from the hub. If this is the
+            // case this integration should not run
+            if (null === $client) {
                 return;
             }
 
-            Hub::getCurrent()->captureException($exception);
+            $integration = $client->getIntegration(self::class);
+
+            // The integration could be binded to a client that is not the one
+            // attached to the current hub. If this is the case, bail out
+            if (null === $integration) {
+                return;
+            }
+
+            $options = $this->options ?? $client->getOptions();
+
+            if (!($options->getErrorTypes() & $exception->getSeverity())) {
+                return;
+            }
+
+            $currentHub->captureException($exception);
         });
     }
 }